Outsourced Degree plan and Course List functionality to their own components, cleaning up App.tsx. Should allow for more flexibility in calling each component.
